requirements.txt
mido==1.2.*
python-rtmidi==1.4.*
pyserial==3.5